Add first two tests for CardPresentPaymentStore
